Implement MVC, add CRUD methods with mongoose
