+++ routes/tourRoutes.js
@@ -5,15 +5,11 @@ const {
   getTour,
   updateTour,
   deleteTour,
-  checkID,
-  checkBody,
 } = require('../controllers/tourController');
 
 const router = express.Router();
 
-router.param('id', checkID);
-
-router.route('/').get(getAllTours).post(checkBody, createTour);
+router.route('/').get(getAllTours).post(createTour);
 
 router.route('/:id').get(getTour).patch(updateTour).delete(deleteTour);
